fixes for moving to fluid depot
