+++ packages/framework/aqueduct/src/helpers/index.ts
@@ -7,3 +7,4 @@ export * from "./primedComponentFactory";
 export * from "./simpleContainerRuntimeFactory";
 export * from "./simpleModuleInstantiationFactory";
 export * from "./sharedComponentFactory";
+export * from "./ComponentClipboardHelper";
