+++ packages/framework/framework-definitions/src/clipboard/IComponentPastable.ts
@@ -2,9 +2,9 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { IComponentContext } from '@microsoft/fluid-runtime-definitions';
+import { IComponentContext } from "@microsoft/fluid-runtime-definitions";
 
-declare module '@microsoft/fluid-component-core-interfaces' {
+declare module "@microsoft/fluid-component-core-interfaces" {
   export interface IComponent extends Readonly<Partial<IProvideComponentPastable>> {}
 }
 
@@ -13,11 +13,15 @@ export interface IProvideComponentPastable {
 }
 
 /**
- * Interface that allows a hosting component to notify a pasted component it is being pasted
+ * Components may implement **IComponentPastable.getComponentUrlOnPaste** to provide an alternate component
+ * identifier to be instantiated during the paste operation. This alternate component should be instantiated
+ * on paste instead of the original component component identifier that was serialized on copy. In essence,
+ * the first instantiated component (serialized component identifier) may act as a factory for the component
+ * that will actually be instantiated.
  */
 export interface IComponentPastable extends IProvideComponentPastable {
   getComponentUrlOnPaste(
     targetContainerContext: IComponentContext,
-    queryParameters: string | undefined
+    queryParameters: string | undefined,
   ): Promise<string | undefined>;
 }
