+++ packages/framework/framework-definitions/src/clipboard/index.ts
@@ -2,6 +2,5 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-export * from './ComponentClipboardHelper';
-export * from './IComponentClipboardProvider';
-export * from './IComponentPastable';
+export * from "./IComponentClipboardProvider";
+export * from "./IComponentPastable";
